requirements.txt
fastapi[standard]>=0.128.0
sqlalchemy>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
bcrypt==4.0.1